Rename some typedefs to avoid name conflicts
C++ does not like that in struct _archiveHandle some of the fields have
the same name as a typedef.  This changes the meaning of the typedef in
the scope of the struct, causing warnings and possibly other problems.

Rename the types so they have names distinct from the struct fields.