Add pg_disable_checksums() and supporting infrastructure
Extracted from a larger patch, this patch provides the basic infrastructure for turning
data checksums off in a cluster.  This also sets up the necessary pg_control fields to
support the necessary multiple states for handling the transitions.

We do a few things:

- Change "data_checksums" from a simple boolean to "data_checksum_state", an enum type
  for all of the potentially-required states for this feature (as well as enabling).

- Add pg_control support for parsing/displaying the new setting.

- Distinguish between the possible checksum states and be specific about whether we care
  about checksum read failures or write failures at all call sites, turning
  DataChecksumsEnabled() into two functions: DataChecksumsEnabledReads() and
  DataChecksumsEnabledWrites().

- Create a superuser function pg_disable_checksums() to perform the actual disabling of
  this in the cluster.

- Add WAL replay/support of the checksum state.

I have *not* changed the default in initdb to enable checksums, but this would be
trivial.