Fix off-by-one in logical slot resource retention
START_REPLICATION ... LOGICAL, the snapshot builder's
SnapBuildXactNeedsSkip(...), etc treat a LSN that exactly equals the start of a
commit record as being inclusive, i.e. we should process that commit and replay
it to the client. If the client does not specify a commit, the start value used
is the replication slot's confirmed_flush lsn, so "confirmed flush" really
means "confirmed flushed up to and not including" - we're replayng the
transaction so we're assuming the client does not already have it.

This is consistent with usage elsewhere, such as walsender's sentPtr meaning
"sent up to but not including this point".

However, LogicalIncreaseXminForSlot(...) and
LogicalIncreaseRestartDecodingForSlot(...) treat confirmed_flush as meaning
"client has flushed everything up to and including this LSN" so they may remove
resources needed by a commit that started at exactly confirmed_flush.

This isn't a big issue in practice. Typically a walsender-using client will
promptly send standby status updates that advance confirmed_flush past the
start-of-commit LSN. The SQL interface advances the replication identifier
to the LSN of end-of-commit + 1, so it is unaffected.