Start sendpos for logical walsender from restart_lsn
In logical streaming mode the walsender would initialize its internal
sentPos to the slot's confirmed_flush position, and the sentPos exposed
to shmem to the slot's restart_lsn. Subseqent updates to both would
use the last-decoded WAL record.

Since we overwrite the internal sentPos before using it, this is clearly an
oversight from an earlier iteration of the logical decoding patch set.
Initialize both to restart_lsn instead.