Drop connection when query can not be cancelled.
While aborting a transaction, if an asynchronous query is running on the
connection, postgres_fdw tries to cancel the query. If the query can not be
cancelled, it means that the connection is broken and sending an ABORT command
would get stuck. This is bad when the transaction is being aborted because user
requested to cancel the query running in the given backend. In such a case,
close the connection instead of trying to "ABORT" the transaction.

Ashutosh Bapat.