Support retaining datadir on successful TAP tests
This moves the datadirs from using tempdirs with randomness in
the directory name to a static name to make it easier to clean
up as well as debug.  The datadir will be retained in case the
the tests fail or the testcode dies/exits with failure, and is
automatically removed on make check.  In case PG_TESTS_NOCLEAN
is defined the datadirs will be retained regardless of test or
exit status.

Additionally, a new make target 'check-clean' is added to the
directories containing TAP tests to allow for cleaning retained
datadirs.