psql: Fix display of whether table is part of publication
If a FOR ALL TABLES publication was present, \d of a table would claim
for each table that it was part of the publication, even for tables that
are ignores, such as system tables and unlogged tables.  Fix the query
by using the function pg_get_publication_tables(), which was intended for
this purpose.

Reported-by: tushar <tushar.ahuja@enterprisedb.com>