Remove endterms and title ids
Older documentation toolchains could not automatically generate text for
xrefs into refentries, so we had to use the endterm attribute to tell it
which text to create.  This is no longer necessary and can be removed.
In most cases, we can just rely on the automatically generated text.  In
the cases where that doesn't work, use the link element instead to
create the desired text locally instead of relying on text at a remote
location.