Add and report the new "session_read_only" GUC pseudo-variable.
Currently, clients wishing to know whether the server is in hot standby
have to resort to polling, which is often suboptimal.

This adds the new "session_read_only" GUC variable that is reported
via a ParameterStatus message.  This allows the clients to:

   (a) know right away that they are connected to a server in hot
       standby; and

   (b) know immediately when a server exits hot standby.

This is immediately useful for target_session_attrs=read-write, as
libpq will not need to make a `SHOW transaction_read_only` roundtrip.
It is also beneficial to various connection pooling systems
(pgpool, pgbouncer etc.)

This also adds support for "target_session_attrs=read-only".

Parts of the patch are based on the work by Tsunakawa Takayuki.