Consistently canonicalize ICU collations' collcollate as BCP 47.
Previously, on versions of ICU prior to ICU 54 collcollate was stored in
the legacy locale format at initdb time.  We now always use the BCP 47
format there.  To make that work, collcollate is now converted from BCP
47 format to the old locale tag format on-the-fly as an ICU collator is
initially opened within a backend (though only on versions of ICU before
ICU 54).  This establishes a general convention: ICU collations'
collcollate should always be in BCP 47 format.

While at it, have CREATE COLLATION perform some simple locale name
sanitization.  This works by assuming that an inability to get a display
name for a given locale name indicates that the locale name is invalid.
The old locale format is now considered to be an implementation detail.
Only BCP 47 was ever documented as supported, so no doc changes are
required here.

In passing, add DEBUG1 message showing the ICU display name of a
collation when a CREATE COLLATION command is executed, and explicitly
name the root locale "root" when populating pg_collation during initdb.