Fix handling of VARIADIC calls for build and array function in json[b]
The following set of functions are declared as VARIADIC ANY from the
start but lacked proper handling for VARIADIC calls:
- json_build_array
- json_build_object
- jsonb_build_array
- jsonb_build_object

In order to address the problem in a common way, introduce in jsonapi.c
a utility function able to take a VARIADIC argument in output to extract
the set of argument values, types and NULL markers which can be then used
by the caller. The function is able to track if a VARIADIC argument is
NULL so as the caller can decide if the result should be an empty object
or not.

Backpatch to 9.4 for the json functions, and 9.5 for the jsonb functions,
where those functions were primarily introduced.

Per bug report from Marko Tiikkaja.

Author: Michael Paquier, Andrew Dunstan
Reviewed-by: Andrew Dunstan, Tom Lane
Discussion: https://postgr.es/m/20171011020058.27986.42308@wrigleys.postgresql.org