Basic implementation of array lists (AList)
ALists do roughly the same job as List but are implemented using Arrays
rather than linked lists. Looping over these lists should be faster than
linked lists and they're particularly useful when the size of the list is
known before any items are added.  Fetching the Nth element in an AList is
an O(1) operations rather than O(n) as it is in the List type.

Array lists not intended as a direct replacement for List. List will still be
useful in cases when good performance is required to delete items out the
middle of a list.

Much is still to do here. This is only the basic implementation. Likely a
more complete version will contain functions for alist_concat and
alist_copy. However, what we have here will be good enough to test the
performance of these vs List for many of today's usages.