polygon-contains-lseg-loop-fix-v00
Fix endless loop on polygon @> polygon

A simple logic flaw on the algorithm was causing polygon contains
polygon operator to loop endlessly.  The algorithm iterates through
the segments of the polygon and checks if they touch or intersect
with each other.  When they do touch or intersect on the inner loop,
the result is found, but the algorithm was continuing to check the rest
of the segments.

The issue reported to the mailing list by Álvaro Herrera, and the test
case is posted by Marco Nenciarini.

Discussion: 20151211174826.GF2762@alvherre.pgsql