Fix ALTER TABLE / ADD COLUMN / IDENTITY
The previous previous coding completely ignored that existing rows for
the new column should have values filled in from the identity sequence.
The previous coding using build_column_default() fails for this because
the sequence ownership isn't registered until after ALTER TABLE, and we
can't do it before because we don't have the column in the catalog yet.
So we specially remember in ColumnDef the sequence name that we decided
on and build a custom NextValueExpr using that.