Fix permissions check on pg_stat_get_wal_senders
Commit 25fff40798fc4ac11a241bfd9ab0c45c085e2212 introduced the
possibility for the pg_read_all_stats to have access to all pg_stat_*
views.
In the discussion, the pg_stat_replication and pg_stat_wal_receiver
views were also considered to be part of that, however
pg_stat_get_wal_senders was somehow not part of that commit, that
seems an oversight.

1: https://www.postgresql.org/message-id/CA%2BOCxoyYxO%2BJmzv2Micj4uAaQdAi6nq0w25BPQgLLxsrvTmREw%40mail.gmail.com