Don't resolve symlinks in find_my_exec()
This reverts 336969e490d71c316a42fabeccda87f798e562dd.  That feature
allowed symlinking a binary to a location outside of the tree and still
be able to find the original installation tree by tracing the symlink.
That setup can still be done by writing a shell script wrapper.

Not resolving symlinks allows installation systems such as Homebrew
where the installation tree is in a version-specific directory but the
publicly visible directory is a symlink.