PQhost to return active connected host and hostaddr details
Earlier PQhost doesn't return the connected host details
when the connection type is CHT_HOST_ADDRESS instead it
returns the provided all connection host parameter values
or the default host details, this can lead to confusion.

It is better to provide the active host or hostaddr details of
the connected server host irrespective of the connection type.
when both host and hostaddr are specified in the connection
string, active host parameter value is returned.

PQhost function returns NULL when the input conn is NULL,
otherwise returns an empty string when the input conn structure
cannot be evaluated.

Similarly PQPort is also changed to return the active connection
port or NULL or an empty string.