Make be-secure-common.c more consistent for future SSL implementations
Recent commit 8a3d9425 which introduced an implementation for SSL
passphrases has introduced at the same time be-secure-common.c which is
aimed at including backend-side APIs which can be used across any
OpenSSL implementations.  The logic behind this file is similar to
fe-secure-openssl.c for the frontend-side APIs.

However, this has forgotten to include check_ssl_key_file_permissions
in the move, which causes a double dependency between be-secure.c and
be-secure-openssl.c.

Refactor the code in a more logic way.  This allows puts into light an
API which is usable by future SSL implementations for permissions on SSL
key files.