Allow functions and procedures with the same name
According to the SQL standard, functions and procedures should have
separate name spaces.  Implement this, by making prokind part of the
primary key of pg_proc.  prokind now only distinguishes functions and
procedures.  Aggregates and window functions are classified under
functions and are distinguished by their old fields proisagg and
proiswindow.  So this partially reverts the changes in
fd1a421fe66173fb9b85d3fe150afde8e812cbe4 that removed those fields.
Various APIs are changed to pass the prokind along with the name and
arguments for a function or procedure lookup.  Conversely, we don't need
to check after the lookup whether we got a function or procedure.

FIXME: regproc and regprocedure currently have no way to pass in
prokind.  They will look for either and fail if the result is not
unique (that is, like a generic "routine" lookup).