Precalculate stable/immutable expressions: infrastructure
Create CacheableExpr which is a generic suberclass for expressions that can be
cacheable. All expression node types that can be cacheable should derive from
CacheableExpr (that is, have CacheableExpr as their first field).  Since
CacheableExpr only contains NodeTag, this is a formality, but it is an easy
form of documentation. In planner if expression can be cached it becomes a part
of CachedExpr node.

To execute cached expressions as PARAM_EXEC, add a list of non-internal cached
expressions in PlannerGlobal / PlannedStmt.

Usually plans for all cached expressions are handled during query planning and
they are compiled separatly. But sometimes they are used in dynamically loaded
plans (for example, domain constraints plans). Usually in this case the new node
PlannedExpr is used.